requirements.txt
BeautifulSoup4
requests
pyyaml
pony
psycopg2cffi
